+++ dummyserver/handlers.py
@@ -1,6 +1,7 @@
 from __future__ import print_function
 
 import collections
+import contextlib
 import gzip
 import json
 import logging
@@ -11,12 +12,8 @@ import zlib
 from io import BytesIO
 from tornado.web import RequestHandler
 
-try:
-    from urllib.parse import urlsplit
-    from http.client import responses
-except ImportError:
-    from urlparse import urlsplit
-    from httplib import responses
+from urllib3.packages.six.moves.http_client import responses
+from urllib3.packages.six.moves.urllib.parse import urlsplit
 
 log = logging.getLogger(__name__)
 
@@ -203,9 +200,8 @@ class TestingApp(RequestHandler):
         if encoding == 'gzip':
             headers = [('Content-Encoding', 'gzip')]
             file_ = BytesIO()
-            zipfile = gzip.GzipFile('', mode='w', fileobj=file_)
-            zipfile.write(data)
-            zipfile.close()
+            with contextlib.closing(gzip.GzipFile('', mode='w', fileobj=file_)) as zipfile:
+                zipfile.write(data)
             data = file_.getvalue()
         elif encoding == 'deflate':
             headers = [('Content-Encoding', 'deflate')]